+++ packages/@tinacms/core/src/cms-forms/form.ts
@@ -18,16 +18,18 @@ limitations under the License.
 
 import arrayMutators from 'final-form-arrays'
 import { FormApi, createForm, Config, Unsubscribe } from 'final-form'
-import { findInactiveFormFields } from './findInactiveFields'
-
-const get = require('lodash.get')
+import { Plugin } from '../plugins'
 
 export interface FormOptions<S> extends Config<S> {
   id: any
   label: string
   fields: Field[]
+  __type?: string
   reset?(): void
   actions?: any[]
+  meta?: {
+    [key: string]: string
+  }
 }
 
 export interface Field {
@@ -53,16 +55,16 @@ interface FieldSubscription {
   unsubscribe: Unsubscribe
 }
 
-export class Form<S = any> {
+export class Form<S = any> implements Plugin {
+  __type: string
   id: any
   label: string
   fields: Field[]
   finalForm: FormApi<S>
   actions: any[]
-  fieldSubscriptions: { [key: string]: FieldSubscription } = {}
-  hiddenFields: { [key: string]: FieldSubscription } = {}
   initialValues: any
   reset?(): void
+  meta: { [key: string]: any }
 
   constructor({
     id,
@@ -73,6 +75,7 @@ export class Form<S = any> {
     ...options
   }: FormOptions<S>) {
     const initialValues = options.initialValues || ({} as S)
+    this.__type = options.__type || 'form'
     this.id = id
     this.label = label
     this.fields = fields
@@ -85,11 +88,19 @@ export class Form<S = any> {
         return response
       },
       mutators: {
-        ...arrayMutators,
+        /**
+         * TODO: Broken by `final-form@4.18.6`.
+         *
+         * Left comment in the docs
+         *
+         * https://github.com/final-form/final-form/pull/275#issuecomment-551132760
+         */
+        ...(arrayMutators as any),
         ...options.mutators,
       },
     })
 
+    this.meta = options.meta || {}
     this.reset = reset
     this.actions = actions || []
     this.initialValues = initialValues
@@ -98,77 +109,6 @@ export class Form<S = any> {
 
   updateFields(fields: Field[]) {
     this.fields = fields
-    /**
-     * We register fields at creation time so we don't have to relay
-     * on `react-final-form` components being rendered.
-     */
-    this.registerFields(this.fields)
-    this.discoverHiddenFields(this.initialValues)
-    this.removeDeclaredFieldsFromHiddenLookup()
-  }
-
-  private discoverHiddenFields(values: any, prefix?: string) {
-    Object.entries(values).map(([name, value]) => {
-      const path = prefix ? `${prefix}.${name}` : name
-
-      if (Array.isArray(value)) {
-        if (value.find(item => typeof item === 'object')) {
-          const prefix = `${path}.INDEX`
-          value.forEach(item => {
-            this.discoverHiddenFields(item, prefix)
-          })
-        } else {
-          this.hiddenFields[path] = {
-            path,
-            field: { name: path, component: null },
-            unsubscribe: this.finalForm.registerField(path, () => {}, {}),
-          }
-        }
-      } else if (typeof value === 'object' && value !== null) {
-        this.discoverHiddenFields(value, path)
-      } else {
-        // Base Case
-        this.hiddenFields[path] = {
-          path,
-          field: { name: path, component: null },
-          unsubscribe: this.finalForm.registerField(path, () => {}, {}),
-        }
-      }
-    })
-  }
-
-  private removeDeclaredFieldsFromHiddenLookup() {
-    Object.keys(this.fieldSubscriptions).forEach(path => {
-      const hiddenField = this.hiddenFields[path]
-      if (hiddenField) {
-        hiddenField.unsubscribe()
-        delete this.hiddenFields[path]
-      }
-    })
-  }
-
-  private registerFields(fields: Field[], pathPrefix?: string) {
-    fields.forEach(field => {
-      const path = pathPrefix ? `${pathPrefix}.${field.name}` : field.name
-
-      const isGroup = ['group'].includes(field.component as string)
-      const isArray = ['group-list', 'blocks'].includes(
-        field.component as string
-      )
-      if (isArray) {
-        const subfields = field.fields || []
-        this.registerFields(subfields, `${path}.INDEX`)
-      } else if (isGroup) {
-        const subfields = field.fields || []
-        this.registerFields(subfields, path)
-      } else {
-        this.fieldSubscriptions[path] = {
-          path,
-          field,
-          unsubscribe: this.finalForm.registerField(path, () => {}, {}),
-        }
-      }
-    })
   }
 
   subscribe: FormApi<S>['subscribe'] = (cb, options) => {
@@ -183,15 +123,43 @@ export class Form<S = any> {
     return this.finalForm.getState().values
   }
 
-  private get inactiveFields() {
-    return findInactiveFormFields(this)
+  get name() {
+    return this.id
   }
 
   updateValues(values: S) {
     this.finalForm.batch(() => {
-      this.inactiveFields.forEach(path => {
-        this.finalForm.change(path, get(values, path))
-      })
+      const activePath: string | undefined = this.finalForm.getState().active
+
+      if (!activePath) {
+        updateEverything(this.finalForm, values)
+      } else {
+        updateSelectively(this.finalForm, values)
+      }
     })
   }
+}
+
+function updateEverything(form: FormApi<any>, values: any) {
+  Object.entries(values).forEach(([path, value]) => {
+    form.change(path, value)
+  })
+}
+
+function updateSelectively(form: FormApi<any>, values: any, prefix?: string) {
+  const activePath: string = form.getState().active!
+
+  Object.entries(values).forEach(([name, value]) => {
+    const path = prefix ? `${prefix}.${name}` : name
+
+    if (typeof value === 'object') {
+      if (activePath.startsWith(path)) {
+        updateSelectively(form, value, path)
+      } else {
+        form.change(path, value)
+      }
+    } else if (path !== activePath) {
+      form.change(path, value)
+    }
+  })
 }