+++ packages/gatsby-tinacms-remark/src/form-actions/delete-action.tsx
@@ -17,18 +17,30 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { CMS } from '@tinacms/core'
+import { Form } from '@tinacms/core'
+import { useCMS } from 'react-tinacms'
+import { ActionButton } from 'tinacms'
 
-export const ERROR_MISSING_CMS = `useCMS could not find an instance of CMS`
+export function DeleteAction({ form }: { form: Form }) {
+  const cms = useCMS()
+  return (
+    <ActionButton
+      onClick={async () => {
+        if (
+          !confirm(
+            `Are you sure you want to delete ${form.values.fileRelativePath}?`
+          )
+        ) {
+          return
+        }
+        await cms.api.git.onDelete!({
+          relPath: form.values.fileRelativePath,
+        })
 
-export const CMSContext = React.createContext<CMS | null>(null)
-
-export function useCMS(): CMS {
-  const cms = React.useContext(CMSContext)
-
-  if (!cms) {
-    throw new Error(ERROR_MISSING_CMS)
-  }
-
-  return cms
+        window.history.back()
+      }}
+    >
+      Delete
+    </ActionButton>
+  )
 }