+++ packages/react-tinacms/src/use-plugin.tsx
@@ -31,10 +31,10 @@ export const usePlugin = usePlugins
  *
  * @alias usePlugin
  */
-export function usePlugins(plugins: Plugin | Plugin[]) {
+export function usePlugins(plugins?: Plugin | Plugin[]) {
   const cms = useCMS()
 
-  let pluginArray: Plugin[]
+  let pluginArray: (Plugin | undefined)[]
 
   if (Array.isArray(plugins)) {
     pluginArray = plugins
@@ -44,12 +44,16 @@ export function usePlugins(plugins: Plugin | Plugin[]) {
 
   React.useEffect(() => {
     pluginArray.forEach(plugin => {
-      cms.plugins.add(plugin)
+      if (plugin) {
+        cms.plugins.add(plugin);
+      }
     })
 
     return () => {
       pluginArray.forEach(plugin => {
-        cms.plugins.remove(plugin)
+        if (plugin) {
+          cms.plugins.remove(plugin);
+        }
       })
     }
   }, [cms.plugins, ...pluginArray])