+++ packages/tinacms/src/components/ActionsMenu.tsx
@@ -20,16 +20,17 @@ import * as React from 'react'
 import styled, { css, StyledComponent } from 'styled-components'
 import { EllipsisVerticalIcon } from '@tinacms/icons'
 import { padding, color, radius, font, shadow } from '@tinacms/styles'
-import '@tinacms/fields/node_modules/@tinacms/styles'
 import { useState, FC } from 'react'
 import { Dismissible } from 'react-dismissible'
 import { useFrameContext } from './SyledFrame'
+import { Form } from '@tinacms/core'
 
 export interface ActionsMenuProps {
+  form: Form
   actions: any[]
 }
 
-export const ActionsMenu: FC<ActionsMenuProps> = ({ actions }) => {
+export const ActionsMenu: FC<ActionsMenuProps> = ({ actions, form }) => {
   const frame = useFrameContext()
   const [actionMenuVisibility, setActionMenuVisibility] = useState(false)
   return (
@@ -46,7 +47,7 @@ export const ActionsMenu: FC<ActionsMenuProps> = ({ actions }) => {
           }}
         >
           {actions.map(Action => (
-            <Action />
+            <Action form={form} />
           ))}
         </Dismissible>
       </ActionsOverlay>