+++ packages/tinacms/src/components/FormView.tsx
@@ -19,7 +19,6 @@ limitations under the License.
 import * as React from 'react'
 import { FormBuilder, FieldsBuilder } from '@tinacms/form-builder'
 
-import '@tinacms/fields/node_modules/@tinacms/styles'
 import { useCMS, useSubscribable } from 'react-tinacms'
 import { useState } from 'react'
 import { Form } from '@tinacms/core'
@@ -37,6 +36,7 @@ import { ActionsMenu } from './ActionsMenu'
 import FormsList from './FormsList'
 import { DragDropContext, DropResult } from 'react-beautiful-dnd'
 import { LeftArrowIcon } from '@tinacms/icons'
+import { LoadingDots } from './LoadingDots'
 import { ResetForm } from './ResetForm'
 import { FORM_HEADER_HEIGHT, SIDEBAR_HEADER_HEIGHT } from '../Globals'
 import { GroupPanel } from '../plugins/fields'
@@ -56,19 +56,9 @@ export const FormsView = () => {
 
   const forms = cms.forms.all()
   const isMultiform = forms.length > 1
-  const activeForm = activeFormId ? cms.forms.findForm(activeFormId) : null
+  const activeForm = activeFormId ? cms.forms.find(activeFormId) : null
 
   const isEditing = !!activeForm
-
-  const moveArrayItem = React.useCallback(
-    (result: DropResult) => {
-      const form = activeForm!.finalForm
-      if (!result.destination) return
-      const name = result.type
-      form.mutators.move(name, result.source.index, result.destination.index)
-    },
-    [activeForm]
-  )
 
   /**
    * No Forms
@@ -88,18 +78,49 @@ export const FormsView = () => {
   }
 
   return (
+    <FormWrapper isEditing={isEditing} isMultiform={isMultiform}>
+      <FormView
+        activeForm={activeForm}
+        setActiveFormId={setActiveFormId}
+        isMultiform={isMultiform}
+      />
+    </FormWrapper>
+  )
+}
+
+export interface FormViewProps {
+  activeForm: Form
+  setActiveFormId(id: string): void
+  isMultiform: boolean
+}
+export function FormView({
+  activeForm,
+  setActiveFormId,
+  isMultiform,
+}: FormViewProps) {
+  const moveArrayItem = React.useCallback(
+    (result: DropResult) => {
+      const form = activeForm!.finalForm
+      if (!result.destination) return
+      const name = result.type
+      form.mutators.move(name, result.source.index, result.destination.index)
+    },
+    [activeForm]
+  )
+
+  return (
     <FormBuilder form={activeForm as any}>
-      {({ handleSubmit, pristine, form }) => {
+      {({ handleSubmit, pristine, form, submitting }) => {
         return (
           <DragDropContext onDragEnd={moveArrayItem}>
-            <FormWrapper isEditing={isEditing} isMultiform={isMultiform}>
-              {isMultiform && (
-                <FormHeader
-                  activeForm={activeForm}
-                  setActiveFormId={setActiveFormId}
-                />
-              )}
-              <FormBody>
+            {isMultiform && (
+              <FormHeader
+                activeForm={activeForm}
+                setActiveFormId={setActiveFormId}
+              />
+            )}
+            <FormBody>
+              <Wrapper>
                 {activeForm &&
                   (activeForm.fields.length ? (
                     <FieldsBuilder
@@ -109,8 +130,10 @@ export const FormsView = () => {
                   ) : (
                     <NoFieldsPlaceholder />
                   ))}
-              </FormBody>
-              <FormFooter>
+              </Wrapper>
+            </FormBody>
+            <FormFooter>
+              <Wrapper>
                 {activeForm.reset && (
                   <ResetForm
                     pristine={pristine}
@@ -123,17 +146,19 @@ export const FormsView = () => {
                 <Button
                   onClick={() => handleSubmit()}
                   disabled={pristine}
+                  busy={submitting}
                   primary
                   grow
                   margin
                 >
-                  Save
+                  {submitting && <LoadingDots />}
+                  {!submitting && 'Save'}
                 </Button>
                 {activeForm.actions.length > 0 && (
-                  <ActionsMenu actions={activeForm.actions} />
+                  <ActionsMenu actions={activeForm.actions} form={activeForm} />
                 )}
-              </FormFooter>
-            </FormWrapper>
+              </Wrapper>
+            </FormFooter>
           </DragDropContext>
         )
       }}
@@ -318,6 +343,13 @@ const FormHeader = styled(
   }
 `
 
+export const Wrapper = styled.div`
+  display: block;
+  margin: 0 auto;
+  max-width: 500px;
+  width: 100%;
+`
+
 export const FormBody: StyledComponent<'div', {}, {}> = styled.div`
   position: relative;
   flex: 1 1 auto;
@@ -328,19 +360,30 @@ export const FormBody: StyledComponent<'div', {}, {}> = styled.div`
   overflow: hidden;
   border-top: 1px solid ${color.grey(2)};
   background-color: #f6f6f9;
+
+  ${Wrapper} {
+    height: 100%;
+    scrollbar-width: none;
+  }
 `
 
 const FormFooter = styled.div`
   position: relative;
   flex: 0 0 auto;
   display: flex;
-  justify-content: space-between;
   align-items: center;
   width: 100%;
   height: 4rem;
   background-color: white;
   border-top: 1px solid ${color.grey(2)};
-  padding: 0 1rem;
+
+  ${Wrapper} {
+    flex: 1 0 auto;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 1rem;
+  }
 `
 
 const FormAnimationKeyframes = keyframes`
@@ -392,7 +435,9 @@ const FormWrapper = styled.div<FormWrapperProps>`
     `};
 `
 
-export const SaveButton: StyledComponent<typeof Button, {}, {}> = styled(Button)`
+export const SaveButton: StyledComponent<typeof Button, {}, {}> = styled(
+  Button
+)`
   flex: 1.5 0 auto;
   padding: 0.75rem 1.5rem;
 `