+++ packages/react-dismissible/CHANGELOG.md
@@ -3,7 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.11](https://github.com/tinacms/tinacms/compare/react-dismissible@1.0.11-alpha.0...react-dismissible@1.0.11) (2019-10-28)
+## [1.0.16](https://github.com/tinacms/tinacms/compare/react-dismissible@1.0.16-alpha.1...react-dismissible@1.0.16) (2019-11-18)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
+## [1.0.16-alpha.1](https://github.com/tinacms/tinacms/compare/react-dismissible@1.0.15...react-dismissible@1.0.16-alpha.1) (2019-11-18)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
+## [1.0.16-alpha.0](https://github.com/tinacms/tinacms/compare/react-dismissible@1.0.15...react-dismissible@1.0.16-alpha.0) (2019-11-18)
 
 **Note:** Version bump only for package react-dismissible